requirements.txt
annotated-types==0.6.0
anyio==4.2.0
appdirs==1.4.4
build==1.0.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
exceptiongroup==1.2.0
fastapi==0.109.2
fsspec==2024.2.0
h11==0.14.0
httpcore==1.0.3
httpx==0.26.0
humanize==4.9.0
idna==3.6
importlib-metadata==7.0.1
importlib-resources==6.1.1
Jinja2==3.1.3
joblib==1.3.2
MarkupSafe==2.1.5
nest-asyncio==1.6.0
numpy==1.26.4
packaging==23.2
palmerpenguins==0.1.4
pandas==2.2.0
pins==0.8.4
pip-tools==7.4.0
plotly==5.19.0
pydantic==2.6.1
pydantic_core==2.16.2
PyJWT==2.8.0
pyproject_hooks==1.0.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
rsconnect_python==1.22.0
scikit-learn==1.4.1.post1
scipy==1.12.0
semver==2.13.0
six==1.16.0
sniffio==1.3.0
starlette==0.36.3
tenacity==8.2.3
threadpoolctl==3.3.0
tomli==2.0.1
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.0
uvicorn==0.27.1
vetiver==0.2.4
xxhash==3.4.1
zipp==3.17.0
